+++ libdariadb/storage/chunk_container.h
@@ -10,7 +10,7 @@ namespace storage {
 
 struct ChunkLink {
   uint64_t id;
-  size_t id_bloom;
+  uint64_t id_bloom;
   dariadb::Time maxTime;
   std::string page_name;
   uint32_t pos;
