+++ libdariadb/storage/index.h
@@ -18,8 +18,8 @@ struct IndexHeader {
   uint32_t chunk_size;        // each chunks size in bytes
   bool is_sorted;             // items in index file sorted by time
   bool is_closed;
-  size_t id_bloom;   // bloom filter of Meas.id
-  size_t flag_bloom; // bloom filter of Meas.flag
+  uint64_t id_bloom;   // bloom filter of Meas.id
+  uint64_t flag_bloom; // bloom filter of Meas.flag
 };
 
 struct IndexReccord {
